stress-membarrier: ensure the cmd type to membarrier call is int
Clean up implicit casting warning by casting cmd to int.

+++ stress-membarrier.c
@@ -69,11 +69,11 @@ static int stress_membarrier_exercise(const stress_args_t *args)
 	mask = (unsigned int)ret;
 	for (i = 1; i; i <<= 1) {
 		if (i & mask) {
-			ret = shim_membarrier(i, 0);
+			ret = shim_membarrier((int)i, 0);
 			(void)ret;
 
 			/* Exercise illegal flags */
-			ret = shim_membarrier(i, ~0);
+			ret = shim_membarrier((int)i, ~0);
 			(void)ret;
 		}
 	}
@@ -81,7 +81,7 @@ static int stress_membarrier_exercise(const stress_args_t *args)
 	/* Exercise illegal command */
 	for (i = 1; i; i <<= 1) {
 		if (!(i & mask)) {
-			ret = shim_membarrier(i, 0);
+			ret = shim_membarrier((int)i, 0);
 			(void)ret;
 			break;
 		}
